requirements.txt
requests>=2.27.1
loguru>=0.6.0
tqdm>=4.62.3
numpy>=1.22.3
matplotlib>=3.5.1
tensorflow>=2.7.0
tensorflow_hub>=0.12.0
python-dotenv>=0.19.2
opencv-python>=4.5.5.64
pillow>=9.0.1
pandas>=1.4.1
